feat(message): 实现消息解析和文件管理功能
- 在MessageHandler中集成GroupMessagesService和PrivateMessagesService进行消息存储
- 添加MessageParser工具类实现消息段解析功能，支持文本、@、表情、文件、图片等多种类型
- 将GroupMessageEvent和PrivateMessageEvent扩展添加parsedMessage字段
- 修改数据库实体类将ID、botId、groupId、senderId、messageTime字段从Integer改为Long类型
- 更新MyBatis映射XML文件适配Long类型字段
- 新增Files数据表和相关实体类、服务接口及实现
- 添加文件管理的DTO类和数据库操作映射文件
- 在消息处理流程中集成消息解析和存储功能

+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/model/domain/GroupMessages.java
@@ -16,17 +16,17 @@ public class GroupMessages implements Serializable {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 
      */
-    private Integer botId;
+    private Long botId;
 
     /**
      * 
      */
-    private Integer groupId;
+    private Long groupId;
 
     /**
      * 
@@ -36,7 +36,7 @@ public class GroupMessages implements Serializable {
     /**
      * 
      */
-    private Integer senderId;
+    private Long senderId;
 
     /**
      * 
@@ -46,7 +46,7 @@ public class GroupMessages implements Serializable {
     /**
      * 
      */
-    private Integer messageTime;
+    private Long messageTime;
 
     @TableLogic
     private Integer isDelete;
