+++ MangoBot-cli/src/main/java/io/github/mangomaner/mangobot/model/domain/PrivateMessages.java
@@ -16,17 +16,17 @@ public class PrivateMessages implements Serializable {
      * 
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 
      */
-    private Integer botId;
+    private Long botId;
 
     /**
      * 
      */
-    private Integer friendId;
+    private Long friendId;
 
     /**
      * 
@@ -36,7 +36,7 @@ public class PrivateMessages implements Serializable {
     /**
      * 
      */
-    private Integer senderId;
+    private Long senderId;
 
     /**
      * 
@@ -46,7 +46,7 @@ public class PrivateMessages implements Serializable {
     /**
      * 
      */
-    private Integer messageTime;
+    private Long messageTime;
 
     @TableLogic
     private Integer isDelete;
